requirements.txt
fastapi~=0.110.1
SQLAlchemy~=2.0.29
greenlet~=3.0.3
python-dotenv~=1.0.1
pandas~=2.2.2
alembic~=1.13.1
mysqlclient